# apitax/ah/api/swagger/apitax/ah/api/models/user_auth.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from apitax.ah.api.models.base_model_ import Model
from apitax.ah.api import util


class UserAuth(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, username=None, password=None, token=None):  # noqa: E501
        """UserAuth - a model defined in Swagger

        :param username: The username of this UserAuth.  # noqa: E501
        :type username: str
        :param password: The password of this UserAuth.  # noqa: E501
        :type password: str
        :param token: The token of this UserAuth.  # noqa: E501
        :type token: str
        """
        self.swagger_types = {
            'username': str,
            'password': str,
            'token': str
        }

        self.attribute_map = {
            'username': 'username',
            'password': 'password',
            'token': 'token'
        }

        self._username = username
        self._password = password
        self._token = token

    @classmethod
    def from_dict(cls, dikt):
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserAuth of this UserAuth.  # noqa: E501
        :rtype: UserAuth
        """
        return util.deserialize_model(dikt, cls)

    @property
    def username(self):
        """Gets the username of this UserAuth.


        :return: The username of this UserAuth.
        :rtype: str
        """
        return self._username

    @username.setter
    def username(self, username):
        """Sets the username of this UserAuth.


        :param username: The username of this UserAuth.
        :type username: str
        """

        self._username = username

    @property
    def password(self):
        """Gets the password of this UserAuth.


        :return: The password of this UserAuth.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password):
        """Sets the password of this UserAuth.


        :param password: The password of this UserAuth.
        :type password: str
        """

        self._password = password

    @property
    def token(self):
        """Gets the token of this UserAuth.


        :return: The token of this UserAuth.
        :rtype: str
        """
        return self._token

    @token.setter
    def token(self, token):
        """Sets the token of this UserAuth.


        :param token: The token of this UserAuth.
        :type token: str
        """

        self._token = token